fix: address code review — reentrancy, print guards, task id, doc, types
- SCAestheticRule: snapshot smoothedScore before first await; single write
  after all async work eliminates partial-update hazard under concurrent callers
- HomeListingAestheticModel: wrap debug print statements in #if DEBUG
- ShotListView: change .task id from imageData.count (collision-prone) to
  imageData (Data is Hashable — guaranteed unique per photo)
- SCRuleResult: update numericScore doc comment to include SCAestheticRule
- SCInstagrammabilityRule.lightingScore: explicit Double() conversion at
  top of function; remove redundant Double() casts on return expressions

+++ Sources/ShotCoachCore/Rules/SCInstagrammabilityRule.swift
@@ -215,11 +215,11 @@ public struct SCInstagrammabilityRule: SCFrameRule {
     /// Narrower than `SCBrightnessRule`'s pass band — rewards well-exposed scenes
     /// without giving a free pass to any non-dark, non-blown frame.
     private func lightingScore(pixelBuffer: CVPixelBuffer) -> Double {
-        let luma = averageLuminance(of: pixelBuffer)
+        let luma = Double(averageLuminance(of: pixelBuffer))
         if luma >= 0.35, luma <= 0.65 { return 1.0 }
-        if luma < 0.35 { return Double(luma / 0.35) }
+        if luma < 0.35 { return luma / 0.35 }
         // luma > 0.65
-        return Double((1.0 - luma) / 0.35)
+        return (1.0 - luma) / 0.35
     }
 
     private func averageLuminance(of pixelBuffer: CVPixelBuffer) -> Float {
